moved quick start
