+++ examples/_quick_start/simple_ocr.py
