+++ examples/_quick_start/summarize_struct.py
