--- quick_start/pipelines/__init__.py
@@ -1,1 +1,0 @@
-
